Quick Settings: add option to turn off main tiles [1/2]
I removed the "dual" status that was hardcoded into Wifi and BT tiles and added it
as a variable which is true when the tile is on the first row.

Also, the first row is now hardcoded to have only 2 items to follow the same design as before.

Inversion and Hotspot tile are no longer contextual and can be added, removed or rearranged at will.

Change-Id: I021e01d436d0a557cc3044ed3d0e17efd95bdf38
Settings: http://review.cyanogenmod.org/83418