Optional allowment of lights
Implement a setting allowing lights to be activated
for new notifications even if the screen is on.

Lights with screen on and Custom values are separated
in an advanced section for a cleaner overview.

This setting gives the user an oportunity to activate lights
with the screen on and also during DayDream screensaver.
The option is not activated by default.

Changes include :
  frameworks/base
  packages/Apps/Settings

Screenshot of the Settings :
  http://i1285.photobucket.com/albums/a583/adriandc/Screenshot_2015-07-20-23-47-13_zpstpmemwwn.png~original

Signed-off-by: AdrianDC <radian.dc@gmail.com>

Lights with screen on: Don't disable leds after the lockscreen

- If lights with screen on is enabled, do not disable leds
  when user passes through the lockscreen.

Change-Id: Ieca282bfc498fb5aa97d44b625e00ad2519a780d
Signed-off-by: AdrianDC <radian.dc@gmail.com>