base: start nfc service prior to systemui
Since registration of NfcService as a manager service happens in the first run of the Nfc app,
and since the SystemUi makes use of this manager service through the NfcTile, NfcManager
can lead into a null reference for the NfcAdapter because the NfcService isn't registered
when the SystemUi starts (this condition depends on the device speed to bring up the call
to the apps).
This change adds a noop service that allow to be called by SystemServer before start the
systemui and ensure that NfcService is available before apps start using NfcManager.

Depends on: http://review.cyanogenmod.org/#/c/108684/
Additional info: http://review.cyanogenmod.org/#/c/108411

Change-Id: I6d36e1ed28af5ea306f7b8c1b15ab49d23e7290d
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>