SystemUI: Enable three icon switching within QS DND tile
This expands the QS DND tile to use three states instead of two.

Java commits by @nicholaschum

What this does:
- Removes the two states (true/false) clause {? XXX:YYY;} and uses an
switch(mZen) instead for the two cases and the default. If in the
future, we want to add another tile, just add another case.
- Adds two vector drawables originally by Cataclysm dev @atl4ntis,
modified and reformatted - this allows for themers to theme this
resource as well by adding the two new drawables.

Change-Id: I339e9e1cfc7b142179822d8ca0aeec254515f66c
PS1: Uses if/elsif/else clause
PS2: Update to use switch(mZen) clause