SystemUI: Fix BT icon duplicate issue
The default code will add new icon to the beginning of the icon list,
this will change the index of the existed icon which will make the
system think it to be new and thus resulting duplicate icon issue.

Change to add the new icon to the end of the list.

Change-Id: I0581242bbf617ee0be3e6d3d664e9b870ee24521
CRs-Fixed: 1046294