SystemUI: Fix BT icon duplicate issue
The default code will add new icon to the beginning of the icon list,
this will change the index of the existed icon which will make the
system think it to be new and thus resulting duplicate icon issue.

Change to add the new icon to the end of the list.

Change-Id: I52873f56a288c32f40dede2e2a44243a65978295
CRs-Fixed: 1046294