Fix split-screen flash issue
For smaller size screens (such as 720p), in split-screen view,
the size of DividerSnapAlgorithm.mTargets is 3, the divider line
can only stop in the middle.

As the original logic, when tapping recent, it will cause the top-docked
activity to resize, leading to flash in some cases.

To fix this, do not call startDragging in this scenario.

Google issue:
https://code.google.com/p/android/issues/detail?id=224851

Change-Id: I3ff051407cac17e770c99d905757d6723f0fc6ac