Allow SetupWizard to do WiFi scans if location off
Location permission is needed for apps to do WiFi scans
and if location is globally off no app can scan too. It
is possible that an OEM does not want location to be on
until the user agrees to that or a device owner can turn
location off before setup. As a result setup wizard
cannot show the WiFi networks preventing the user form
logging in. This change allows the setup wizard component
to perform WiFi scans even if location is globally off.

bug:29577364
Change-Id: Ifb5542266f22d121e5d4f31ee7a751c3365b80bf