Fix Bluetooth icon not properly refreshing
In the previous commit, a ContentObserver was being used and wasn't picking up
on the changes to the preference until the notification shade was toggled.

Borrowing heavily from the logic of c4d4dc585cc1581ce656b22ead3352a28609a8ae,
using a SettingsObserver fixes this and allows the toggle to happen in realtime.

Change-Id: I71b6c87d48bdbd56a90913ef1b7b5ccf9b7316f9
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>