Fix Bluetooth icon not properly refreshing
In the previous commit, a ContentObserver was being used and wasn't picking up
on the changes to the preference until the notification shade was toggled.

Borrowing heavily from the logic of c4d4dc585cc1581ce656b22ead3352a28609a8ae,
using a SettingsObserver fixes this and allows the toggle to happen in realtime.

Change-Id: I9f3d8187197271a68fc706fb4190e6b0e65dc77f
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>