Fix memory corruption caused by patchCache.clear
The workfolw of drawPatch in HWUI is like this:
applyDraw -> getMesh -> patchCache.get -> GlopBuilder -> renderGlop
-> RenderState.render

If trimMemory is invoked before drawPatch, the content in mesh may
be corrupted by patchCache.clear. Meanwhile, mGenerationId is still
the same before and after clear operation, so mesh is not updated,
the wrong info is used in rendering, thus lead to driver crash.

Test: run com.immomo.momo app which reproduce the crash

Change-Id: I1ae9b6e0c79f6783cc72a2730bfd112dce60bab0
CRs-fixed: 1088161